requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
nltk
sentence-transformers
torch
scikit-learn
tqdm
websockets
groq